Fixed some layer args being set to int64 which cannot be dumped to JSON

+++ Core/Layer.py
@@ -235,36 +235,36 @@ class ArgRegularizers(EnumWithNone):
 
 
 def gen_2d_kernel_size(input_size):
-    kernel_size = random.choice(range(1, input_size, 2))
+    kernel_size = int(random.choice(range(1, input_size, 2)))
     return (kernel_size, kernel_size)
 
 
 def gen_3d_strides(max_bound):
-    stride_size = random.randint(1, max_bound)
+    stride_size = int(random.randint(1, max_bound))
     return (stride_size, stride_size, stride_size)
 
 
 def gen_2d_strides(max_bound):
-    stride_size = random.randint(1, max_bound)
+    stride_size = int(random.randint(1, max_bound))
     return (stride_size, stride_size)
 
 
 def gen_1d_strides(max_bound):
-    return random.randint(1, max_bound)
+    return int(random.randint(1, max_bound))
 
 
 def gen_3d_poolsize(max_bound):
-    size = random.randint(1, max_bound)
+    size = int(random.randint(1, max_bound))
     return (size, size, size)
 
 
 def gen_2d_poolsize(max_bound):
-    size = random.randint(1, max_bound)
+    size = int(random.randint(1, max_bound))
     return (size, size)
 
 
 def gen_1d_poolsize(max_bound):
-    return random.randint(1, max_bound)
+    return int(random.randint(1, max_bound))
 
 
 def gen_2d_dilation():
@@ -274,7 +274,7 @@ def gen_2d_dilation():
 
 def gen_dropout(max):
     while True:
-        ret = round(random.uniform(0, max), 2)
+        ret = int(round(random.uniform(0, max), 2))
         if ret != 0.0:
             return ret
